requirements.txt
mysql-connector-python
PyQt6
